--- src/main/java/dongheon/lotto/domain/WinningLotto.java
@@ -1,25 +1,0 @@
-package dongheon.lotto.domain;
-
-import dongheon.lotto.view.InputView;
-
-public class WinningLotto {
-    private final Lotto lotto;
-    private final int bonus;
-
-    private WinningLotto(Lotto lotto, int bonus) {
-        this.lotto = lotto;
-        this.bonus = bonus;
-    }
-
-    public Rank match(Lotto userLotto) {
-        int countOfMatch = userLotto.countOfMatch(lotto);
-        boolean matchBonus = userLotto.containNumber(bonus);
-        return Rank.valueOf(countOfMatch, matchBonus);
-    }
-
-    public static WinningLotto getWinningLotto() {
-        Lotto lastWeekLotto = InputView.getLastWeekLotto();
-        int bonus = InputView.getLastWeekBonus(lastWeekLotto.getNumbers());
-        return new WinningLotto(lastWeekLotto, bonus);
-    }
-}
