--- src/main/java/dongheon/lotto/interfaces/LottoCreateStrategy.java
@@ -1,9 +1,0 @@
-package dongheon.lotto.interfaces;
-
-import dongheon.lotto.domain.LottoGame;
-
-import java.util.List;
-
-public interface LottoCreateStrategy {
-    void create(LottoGame lottoGame);
-}
