+++ src/main/java/lotto/interfaces/LottoCreateStrategy.java
@@ -1,0 +1,8 @@
+package lotto.interfaces;
+
+import lotto.model.LottoRepository;
+
+
+public interface LottoCreateStrategy {
+    void create(LottoRepository lottoRepository, int countOfLotto);
+}
