+++ src/main/java/lotto/utils/RandomLottoCreator.java
@@ -1,0 +1,22 @@
+package lotto.utils;
+
+import lotto.model.Lotto;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+import static lotto.model.Lotto.*;
+
+public class RandomLottoCreator {
+    private final List<Integer> randomBox = IntStream
+            .rangeClosed(LOTTO_MIN_VALUE, LOTTO_MAX_VALUE)
+            .boxed().collect(Collectors.toList());
+
+    public Lotto createRandomLotto() {
+        Collections.shuffle(randomBox);
+        List<Integer> lottoNum = randomBox.subList(ZERO, LOTTO_LENGTH);
+        return new Lotto(lottoNum);
+    }
+}
