--- src/test/java/dongheon/lotto/domain/LottoGameTest.java
@@ -1,7 +1,0 @@
-package dongheon.lotto.domain;
-
-class LottoGameTest {
-    /*init을 테스트 하고 싶은데 수동 입력이 포함되어 있어서
-            어떻게 테스트해야할 지 모르겠네요 ..*/
-
-}
